+++ api_version.h
@@ -1,0 +1,10 @@
+#ifndef API_VERSION_H_
+#define API_VERSION_H_
+
+extern void set_api_with_seid(bool);
+extern bool get_api_with_seid(void);
+
+extern void set_api_with_urr_bar(bool);
+extern bool get_api_with_urr_bar(void);
+
+#endif
