+++ genl_urr.c
@@ -1,0 +1,28 @@
+#include <net/genetlink.h>
+#include <net/sock.h>
+
+#include "genl_urr.h"
+
+int gtp5g_genl_add_urr(struct sk_buff *skb, struct genl_info *info)
+{
+    /* Not implemented yet */
+    return -1; 
+}
+
+int gtp5g_genl_del_urr(struct sk_buff *skb, struct genl_info *info)
+{
+    /* Not implemented yet */
+    return -1; 
+}
+
+int gtp5g_genl_get_urr(struct sk_buff *skb, struct genl_info *info)
+{
+    /* Not implemented yet */
+    return -1; 
+}
+
+int gtp5g_genl_dump_urr(struct sk_buff *skb, struct netlink_callback *cb)
+{
+    /* Not implemented yet */
+    return -1; 
+}
