+++ hash.c
@@ -1,0 +1,5 @@
+#include <linux/jhash.h>
+
+#include "hash.h"
+
+u32 gtp5g_h_initval;
