+++ util.c
@@ -1,0 +1,11 @@
+#include <linux/kernel.h> 
+
+#include "util.h"
+
+void ip_string(char * ip_str, int ip_int){
+    sprintf(ip_str, "%i.%i.%i.%i",
+          (ip_int) & 0xFF,
+          (ip_int >> 8) & 0xFF,
+          (ip_int >> 16) & 0xFF,
+          (ip_int >> 24) & 0xFF);
+}
