+++ util.h
@@ -1,0 +1,6 @@
+#ifndef UTIL_H_
+#define UTIL_H_
+
+extern void ip_string(char *, int);
+
+#endif
